[fix] Prevent sending multiple identical notifications #277
If a user is in several organizations,
they will receive multiple identical notifications.
The reason for this is that the database query,
which determines which user should receive the notification,
does a LEFT OUTER JOIN with the "openwisp_users_organizationuser" table.
In this table, a user is multiple times listed,
when he is in several organizations.
Therefore the query returns such a user multiple times,
which will then receive multiple identical notifications.
This commit fixes this bug by adding a distinct to all user queries.

Fixes #277

Signed-off-by: Konrad Kreitmair <kkreitmair@tdt.de>